+++ src/app/_shared-components/SetIdentity/SetIdentity.tsx
@@ -4,38 +4,27 @@
 
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useIdentityService } from '@/hooks/useIdentityService';
 import { useUserPreferences } from '@/hooks/useUserPreferences';
-import { Form, FormField, FormItem, FormLabel, FormControl, FormMessage } from '@ui/Form';
 import { useForm } from 'react-hook-form';
 import { useTranslations } from 'next-intl';
-import { getCurrentNetwork } from '@/_shared/_utils/getCurrentNetwork';
-import { ValidatorService } from '@/_shared/_services/validator_service';
 import { BN, BN_ZERO } from '@polkadot/util';
-import { useToast } from '@/hooks/useToast';
-import { ENotificationStatus, IOnChainIdentity } from '@/_shared/types';
 import Link from 'next/link';
 import { useUser } from '@/hooks/useUser';
-import Image from 'next/image';
-import EmailIcon from '@assets/icons/email-icon-dark.svg';
-import TwitterIcon from '@assets/icons/twitter-icon-dark.svg';
-import RiotIcon from '@assets/icons/riot-icon.svg';
 import { NETWORKS_DETAILS } from '@/_shared/_constants/networks';
+import { getCurrentNetwork } from '@/_shared/_utils/getCurrentNetwork';
 import { useQuery } from '@tanstack/react-query';
 import { FIVE_MIN_IN_MILLI } from '@/app/api/_api-constants/timeConstants';
-import { AlertCircle, ArrowLeft } from 'lucide-react';
-import { Separator } from '../Separator';
+import { ArrowLeft } from 'lucide-react';
+import { useSearchParams } from 'next/navigation';
+import { ESetIdentityStep } from '@/_shared/types';
 import { Button } from '../Button';
-import { Input } from '../Input';
 import SetIdentityFees from './SetIdentityFees/SetIdentityFees';
-import SocialVerifications from './SocialVerifications/SocialVerifications';
+import RequestJudgement from './RequestJudgement/RequestJudgement';
 import IdentitySuccessScreen from './IdentitySuccessScreen/IdentitySuccessScreen';
-import SwitchWalletOrAddress from '../SwitchWalletOrAddress/SwitchWalletOrAddress';
-import IdentityFeeCollaps from './IdentityFeeCollaps/IdentityFeeCollaps';
-import AddressRelationsPicker from '../AddressRelationsPicker/AddressRelationsPicker';
-import { Alert, AlertDescription } from '../Alert';
 import TeleportToPeopleChain from '../TeleportFunds/TeleportToPeopleChain';
+import SetIdentityForm from './SetIdentityForm/SetIdentityForm';
 
 interface ISetIdentityFormFields {
 	displayName: string;
@@ -45,44 +34,22 @@ interface ISetIdentityFormFields {
 	matrix?: string;
 }
 
-enum ESetIdentityStep {
-	GAS_FEE,
-	SET_IDENTITY_FORM,
-	IDENTITY_SUCCESS,
-	SOCIAL_VERIFICATION,
-	TELEPORT_TO_PEOPLE_CHAIN
-}
-
-function SocialIcon({ icon }: { icon: string }) {
-	return (
-		<div className='flex h-10 w-10 items-center justify-center rounded-full bg-border_grey/40'>
-			<Image
-				src={icon}
-				alt='social icon'
-				width={24}
-				height={24}
-			/>
-		</div>
-	);
-}
 function SetIdentity() {
 	const t = useTranslations();
 	const { user } = useUser();
 	const { userPreferences } = useUserPreferences();
-	const { toast } = useToast();
 
 	const network = getCurrentNetwork();
 
 	const formData = useForm<ISetIdentityFormFields>();
 
-	const [loading, setLoading] = useState(false);
-	const [identityLoading, setIdentityLoading] = useState(false);
-
 	const { identityService } = useIdentityService();
 
-	const [step, setStep] = useState<ESetIdentityStep>(ESetIdentityStep.GAS_FEE);
+	const searchParams = useSearchParams();
 
-	const [identityValues, setIdentityValues] = useState<IOnChainIdentity>();
+	const open = searchParams.get('open');
+
+	const [step, setStep] = useState<ESetIdentityStep>(open && Object.values(ESetIdentityStep).includes(open) ? (open as unknown as ESetIdentityStep) : ESetIdentityStep.GAS_FEE);
 
 	const fetchRegistrarFees = async () => {
 		if (!identityService) return null;
@@ -104,82 +71,6 @@ function SetIdentity() {
 		refetchOnWindowFocus: false
 	});
 
-	const fetchGasFee = async () => {
-		if (!identityService) return null;
-
-		const gasFee = await identityService.getGasFee({
-			address: userPreferences.selectedAccount?.address || '',
-			registrarFee: registrarFee || BN_ZERO,
-			displayName: formData.getValues('displayName'),
-			email: formData.getValues('email'),
-			legalName: formData.getValues('legalName'),
-			twitter: formData.getValues('twitter'),
-			matrix: formData.getValues('matrix')
-		});
-
-		return gasFee ? new BN(gasFee) : BN_ZERO;
-	};
-	const { data: gasFee } = useQuery({
-		queryKey: ['gasFee', user?.id, userPreferences.selectedAccount?.address, formData.getValues('displayName'), formData.getValues('email')],
-		queryFn: () => fetchGasFee(),
-		placeholderData: (previousData) => previousData,
-		staleTime: FIVE_MIN_IN_MILLI,
-		retry: false
-	});
-
-	useEffect(() => {
-		const setDefaultIdentityValues = async () => {
-			if (!identityService || !network || !userPreferences.selectedAccount?.address) return;
-
-			setIdentityLoading(true);
-			const identityInfo = await identityService.getOnChainIdentity(userPreferences.selectedAccount.address);
-
-			setIdentityValues(identityInfo);
-			formData.setValue('displayName', identityInfo.display);
-			formData.setValue('legalName', identityInfo.legal);
-			formData.setValue('email', identityInfo.email);
-			formData.setValue('twitter', identityInfo.twitter);
-			formData.setValue('matrix', identityInfo.matrix);
-
-			setIdentityLoading(false);
-		};
-		setDefaultIdentityValues();
-	}, [formData, identityService, network, userPreferences.selectedAccount?.address]);
-
-	const handleSetIdentity = async (values: ISetIdentityFormFields) => {
-		if (!userPreferences.wallet || !userPreferences.selectedAccount?.address || !identityService) return;
-
-		const { displayName, legalName, email, twitter, matrix } = values;
-		setLoading(true);
-
-		await identityService.setOnChainIdentity({
-			address: userPreferences.selectedAccount.address,
-			displayName,
-			email,
-			legalName,
-			twitter,
-			matrix,
-			registrarFee: registrarFee || BN_ZERO,
-			onSuccess: () => {
-				setLoading(false);
-				toast({
-					status: ENotificationStatus.SUCCESS,
-					title: t('SetIdentity.success'),
-					description: t('SetIdentity.successDescription')
-				});
-				setStep(ESetIdentityStep.IDENTITY_SUCCESS);
-			},
-			onFailed: (errorMessageFallback?: string) => {
-				setLoading(false);
-				toast({
-					status: ENotificationStatus.ERROR,
-					title: t('SetIdentity.failed'),
-					description: errorMessageFallback || t('SetIdentity.failedDescription')
-				});
-			}
-		});
-	};
-
 	if (!user) {
 		return (
 			<p className='flex items-center gap-x-1 text-center text-sm text-text_primary'>
@@ -198,8 +89,7 @@ function SetIdentity() {
 	return step === ESetIdentityStep.GAS_FEE ? (
 		<SetIdentityFees
 			onNext={() => setStep(ESetIdentityStep.SET_IDENTITY_FORM)}
-			onRequestJudgement={() => setStep(ESetIdentityStep.SOCIAL_VERIFICATION)}
-			disabledRequestJudgement={!identityValues?.display || !identityValues?.email || !identityValues?.hash}
+			onRequestJudgement={() => setStep(ESetIdentityStep.REQUEST_JUDGEMENT)}
 			registrarFee={registrarFee || BN_ZERO}
 		/>
 	) : step === ESetIdentityStep.IDENTITY_SUCCESS && userPreferences.selectedAccount?.address ? (
@@ -210,10 +100,10 @@ function SetIdentity() {
 			legalName={formData.getValues('legalName')}
 			twitter={formData.getValues('twitter')}
 			matrix={formData.getValues('matrix')}
-			onNext={() => setStep(ESetIdentityStep.SOCIAL_VERIFICATION)}
+			onNext={() => setStep(ESetIdentityStep.REQUEST_JUDGEMENT)}
 		/>
-	) : step === ESetIdentityStep.SOCIAL_VERIFICATION ? (
-		<SocialVerifications />
+	) : step === ESetIdentityStep.REQUEST_JUDGEMENT ? (
+		<RequestJudgement onSetIdentity={() => setStep(ESetIdentityStep.SET_IDENTITY_FORM)} />
 	) : step === ESetIdentityStep.TELEPORT_TO_PEOPLE_CHAIN ? (
 		<div className='flex flex-1 flex-col gap-y-4'>
 			<div>
@@ -234,207 +124,11 @@ function SetIdentity() {
 			/>
 		</div>
 	) : (
-		<Form {...formData}>
-			<form
-				className='flex flex-1 flex-col overflow-y-hidden'
-				onSubmit={formData.handleSubmit(handleSetIdentity)}
-			>
-				<div className='flex flex-1 flex-col gap-y-4 overflow-y-auto'>
-					<Alert
-						variant='info'
-						className='flex items-center gap-x-3'
-					>
-						<AlertCircle className='h-4 w-4' />
-						<AlertDescription className='flex w-full items-center justify-between'>
-							<h2 className='text-sm font-medium'>{t('SetIdentity.TeleportFundsToPeopleChain')}</h2>
-							<Button
-								onClick={() => setStep(ESetIdentityStep.TELEPORT_TO_PEOPLE_CHAIN)}
-								size='sm'
-								className='h-6'
-							>
-								{t('SetIdentity.Teleport')}
-							</Button>
-						</AlertDescription>
-					</Alert>
-
-					<SwitchWalletOrAddress
-						small
-						customAddressSelector={
-							<AddressRelationsPicker
-								withBalance
-								showPeopleChainBalance
-							/>
-						}
-						disabled={identityLoading}
-					/>
-
-					{/* Display Name */}
-					<FormField
-						control={formData.control}
-						name='displayName'
-						key='displayName'
-						disabled={loading || identityLoading}
-						rules={{
-							required: true,
-							validate: (value) => {
-								if (value?.length === 0) return 'Invalid Display Name';
-								return true;
-							}
-						}}
-						render={({ field }) => (
-							<FormItem>
-								<FormLabel>{t('SetIdentity.displayName')}</FormLabel>
-								<FormControl>
-									<Input
-										placeholder={t('SetIdentity.displayName')}
-										{...field}
-									/>
-								</FormControl>
-								<FormMessage />
-							</FormItem>
-						)}
-					/>
-
-					{/* Legal Name */}
-					<FormField
-						control={formData.control}
-						name='legalName'
-						key='legalName'
-						disabled={loading}
-						rules={{
-							required: true,
-							validate: (value) => {
-								if (value?.length === 0) return 'Invalid Display Name';
-								return true;
-							}
-						}}
-						render={({ field }) => (
-							<FormItem>
-								<FormLabel>{t('SetIdentity.legalName')}</FormLabel>
-								<FormControl>
-									<Input
-										placeholder={t('SetIdentity.legalName')}
-										{...field}
-									/>
-								</FormControl>
-								<FormMessage />
-							</FormItem>
-						)}
-					/>
-
-					<Separator />
-
-					<p className='font-semibold text-wallet_btn_text'>Socials</p>
-
-					{/* Email */}
-					<FormField
-						control={formData.control}
-						name='email'
-						key='email'
-						disabled={loading}
-						rules={{
-							validate: (value) => {
-								if (!ValidatorService.isValidEmail(value)) return 'Invalid Email';
-								return true;
-							}
-						}}
-						render={({ field }) => (
-							<FormItem className='flex items-center gap-x-2'>
-								<div className='flex w-28 items-center gap-x-2'>
-									<SocialIcon icon={EmailIcon} />
-									<FormLabel>{t('SetIdentity.email')}</FormLabel>
-								</div>
-								<div className='flex-1'>
-									<FormControl>
-										<Input
-											placeholder={t('SetIdentity.emailPlaceholder')}
-											{...field}
-										/>
-									</FormControl>
-									<FormMessage />
-								</div>
-							</FormItem>
-						)}
-					/>
-
-					{/* Twitter */}
-					<FormField
-						control={formData.control}
-						name='twitter'
-						key='twitter'
-						disabled={loading}
-						rules={{
-							validate: (value) => {
-								if (value && !ValidatorService.isValidTwitterHandle(value)) return 'Invalid Twitter Handle';
-								return true;
-							}
-						}}
-						render={({ field }) => (
-							<FormItem className='flex items-center gap-x-2'>
-								<div className='flex w-28 items-center gap-x-2'>
-									<SocialIcon icon={TwitterIcon} />
-									<FormLabel>{t('SetIdentity.twitter')}</FormLabel>
-								</div>
-								<div className='flex-1'>
-									<FormControl>
-										<Input
-											placeholder={t('SetIdentity.twitterPlaceholder')}
-											{...field}
-										/>
-									</FormControl>
-									<FormMessage />
-								</div>
-							</FormItem>
-						)}
-					/>
-
-					{/* Matrix */}
-					<FormField
-						control={formData.control}
-						name='matrix'
-						key='matrix'
-						rules={{
-							validate: (value) => {
-								if (value && !ValidatorService.isValidMatrixHandle(value)) return 'Invalid Matrix Handle';
-								return true;
-							}
-						}}
-						disabled={loading}
-						render={({ field }) => (
-							<FormItem className='flex items-center gap-x-2'>
-								<div className='flex w-28 items-center gap-x-2'>
-									<SocialIcon icon={RiotIcon} />
-									<FormLabel>{t('SetIdentity.riot')}</FormLabel>
-								</div>
-								<div className='flex-1'>
-									<FormControl>
-										<Input
-											placeholder={t('SetIdentity.riotPlaceholder')}
-											{...field}
-										/>
-									</FormControl>
-									<FormMessage />
-								</div>
-							</FormItem>
-						)}
-					/>
-				</div>
-				<IdentityFeeCollaps
-					className='mt-4'
-					registrarFee={registrarFee || BN_ZERO}
-					gasFee={gasFee || BN_ZERO}
-				/>
-				<Separator className='my-4' />
-				<div className='flex justify-end'>
-					<Button
-						isLoading={loading}
-						type='submit'
-					>
-						{t('SetIdentity.setIdentity')}
-					</Button>
-				</div>
-			</form>
-		</Form>
+		<SetIdentityForm
+			registrarFee={registrarFee || BN_ZERO}
+			onTeleport={() => setStep(ESetIdentityStep.TELEPORT_TO_PEOPLE_CHAIN)}
+			onSuccess={() => setStep(ESetIdentityStep.IDENTITY_SUCCESS)}
+		/>
 	);
 }
