+++ src/app/api/_api-utils/getOauthConsumer.ts
@@ -5,7 +5,7 @@ import oauth from 'oauth';
 import { ERROR_CODES } from '@/_shared/_constants/errorLiterals';
 import { StatusCodes } from 'http-status-codes';
 import { ENetwork, ESocial } from '@/_shared/types';
-import { TWITTER_CONSUMER_API_KEY, TWITTER_CONSUMER_API_SECRET_KEY, VERIFICATION_CALLBACK_URL } from '@/app/api/_api-constants/apiEnvVars';
+import { TWITTER_CONSUMER_API_KEY, TWITTER_CONSUMER_API_SECRET_KEY } from '@/app/api/_api-constants/apiEnvVars';
 import { APIError } from './apiError';
 
 const OAUTH_CONSUMER_ERROR = 'TWITTER_CONSUMER_API_KEY or TWITTER_CONSUMER_API_SECRET_KEY missing in env';
@@ -20,7 +20,7 @@ export const getOauthConsumer = (network: ENetwork) => {
 		TWITTER_CONSUMER_API_KEY,
 		TWITTER_CONSUMER_API_SECRET_KEY,
 		'1.0A',
-		`${VERIFICATION_CALLBACK_URL}?social=${ESocial.TWITTER}`,
+		`https://${network}.polkassembly.io/confirm-verification?social=${ESocial.TWITTER}`,
 		'HMAC-SHA1'
 	);
 };
