+++ src/intl/messages/es.json
@@ -802,6 +802,7 @@
 		"legalName": "Nombre legal",
 		"email": "Correo electrónico",
 		"twitter": "Twitter",
+		"socials": "Redes Sociales",
 		"riot": "Riot",
 		"emailPlaceholder": "Ingrese su dirección de correo electrónico",
 		"twitterPlaceholder": "@SuNombreTwitter",
@@ -832,11 +833,15 @@
 		"verifying": "Verificando...",
 		"verificationSuccessful": "Verificación exitosa",
 		"verificationFailed": "Verificación fallida",
-		"goToHome": "Ir al inicio",
+		"goToRequestJudgement": "Ir a Solicitar Juicio",
 		"verificationEmailSent": "Correo de verificación enviado",
 		"verificationEmailSentDescription": "Por favor revise su correo electrónico para la verificación",
 		"emailNotSent": "Correo no enviado",
-		"emailNotSentDescription": "Hubo un error al enviar el correo de verificación"
+		"emailNotSentDescription": "Hubo un error al enviar el correo de verificación",
+		"IdentityNotSet": "Identidad No Establecida",
+		"IdentityNotSetDescription": "Para solicitar un juicio, necesitas tener tu identidad en cadena establecida primero.",
+		"IdentityNotSetDescription2": "El juicio es una forma para que los registradores verifiquen tu perfil y proporcionen un nivel de confiabilidad basado en tu información de identidad (como nombre para mostrar, correo electrónico, sitio web, nombre legal, etc.).",
+		"IdentityNotSetDescription3": "Una vez que hayas establecido tu identidad, puedes solicitar un juicio de un registrador. Este juicio será visible para otros y te ayuda a construir credibilidad en cadena."
 	},
 	"Overview": {
 		"month": "Mes",
