+++ src/intl/messages/ja.json
@@ -801,6 +801,7 @@
 		"legalName": "法的名称",
 		"email": "メールアドレス",
 		"twitter": "Twitter",
+		"socials": "ソーシャルメディア",
 		"riot": "Riot",
 		"emailPlaceholder": "メールアドレスを入力してください",
 		"twitterPlaceholder": "@あなたのTwitter名",
@@ -831,11 +832,15 @@
 		"verifying": "検証中...",
 		"verificationSuccessful": "検証成功",
 		"verificationFailed": "検証失敗",
-		"goToHome": "ホームに戻る",
+		"goToRequestJudgement": "判定リクエストに移動",
 		"verificationEmailSent": "確認メールを送信しました",
 		"verificationEmailSentDescription": "確認のためメールをご確認ください",
 		"emailNotSent": "メール送信失敗",
-		"emailNotSentDescription": "確認メールの送信中にエラーが発生しました"
+		"emailNotSentDescription": "確認メールの送信中にエラーが発生しました",
+		"IdentityNotSet": "アイデンティティが設定されていません",
+		"IdentityNotSetDescription": "判定をリクエストするには、まずオンチェーンアイデンティティを設定する必要があります。",
+		"IdentityNotSetDescription2": "判定は、登録機関があなたのプロフィールを検証し、あなたのアイデンティティ情報（表示名、メール、ウェブサイト、法的名称など）に基づいて信頼性のレベルを提供する方法です。",
+		"IdentityNotSetDescription3": "アイデンティティを設定したら、登録機関に判定をリクエストできます。この判定は他の人に表示され、オンチェーンで信頼性を築くのに役立ちます。"
 	},
 	"Overview": {
 		"month": "月",
