+++ src/intl/messages/zh.json
@@ -801,6 +801,7 @@
 		"legalName": "法定名称",
 		"email": "电子邮箱",
 		"twitter": "Twitter",
+		"socials": "社交媒体",
 		"riot": "Riot",
 		"emailPlaceholder": "请输入您的电子邮箱地址",
 		"twitterPlaceholder": "@您的Twitter用户名",
@@ -831,11 +832,15 @@
 		"verifying": "验证中...",
 		"verificationSuccessful": "验证成功",
 		"verificationFailed": "验证失败",
-		"goToHome": "返回首页",
+		"goToRequestJudgement": "前往请求判断",
 		"verificationEmailSent": "验证邮件已发送",
 		"verificationEmailSentDescription": "请检查您的邮箱进行验证",
 		"emailNotSent": "邮件未发送",
-		"emailNotSentDescription": "发送验证邮件时出现错误"
+		"emailNotSentDescription": "发送验证邮件时出现错误",
+		"IdentityNotSet": "身份未设置",
+		"IdentityNotSetDescription": "要请求判断，您需要首先设置您的链上身份。",
+		"IdentityNotSetDescription2": "判断是注册商验证您的个人资料并根据您的身份信息（如显示名称、电子邮箱、网站、法定名称等）提供可信度级别的方式。",
+		"IdentityNotSetDescription3": "一旦您设置了身份，您就可以向注册商请求判断。这个判断将对其他人可见，并帮助您在链上建立可信度。"
 	},
 	"Bounties": {
 		"onchainBounties": "链上赏金",
